A function to replace a token with a value in a String.
Haskell libraries has a Data.Text.replace already achieves everything that is mentioned in this i.e. replaces a token with a value and works on Data.Text types. This function replaces a token with a value, in the target string.

Example usage is provided below -

Prelude Data.List.Split> repl "@name@" "Thulasiram" "Hello @name@. Welcome to the world of Haskell..."
"Hello Thulasiram. Welcome to the world of Haskell..."